Move all "views" filest to "routes" submodule. Move routes from __init__.py to routes/general.py
