+++ anarchoApp/anarcho/routes/__init__.py
@@ -1,0 +1,3 @@
+"""
+All routes of Flask application
+"""
